Update project.pbxproj
As per the README
```
Unless you are using tvOS, repeat the same steps to add the iAd.framework, CoreTelephony.framework, AppTrackingTransparency.framework and StoreKit.framework. Change the Status of both frameworks to Optional.
```
Not doing this may lead to problems on iOS 12 and 11 due to the framework not being available there.